Switching to docker-compose v2 and some minor fixes for x86_64

+++ docker-compose-tests.sh
@@ -4,12 +4,13 @@ set -e
 COMPOSE_PROJDIR=$(basename $PWD)
 ARROW_MSGBUS=${ARROW_MSGBUS:-kafka}
 
-#docker-compose down --remove-orphans #&& docker rmi ${COMPOSE_PROJDIR}_builder ${COMPOSE_PROJDIR}_stolon || echo new or partial install
+#docker-compose down --remove-orphans #&& docker rmi ${COMPOSE_PROJDIR}-builder ${COMPOSE_PROJDIR}-stolon || echo new or partial install
 docker-compose up -d "${ARROW_MSGBUS}" postgres builder
-docker exec -u postgres ${COMPOSE_PROJDIR}_postgres_1 /usr/bin/psql -tc "select datname from pg_database where datname='src'" | grep -q src || docker exec -u postgres ${COMPOSE_PROJDIR}_postgres_1 /usr/bin/psql -c '\i /host/config/schema.sql'
-docker exec -ti ${COMPOSE_PROJDIR}_builder_1 /bin/bash -ic "cd /host && make build"
+docker exec -u postgres ${COMPOSE_PROJDIR}-postgres-1 /usr/bin/psql -tc "select datname from pg_database where datname='src'" | grep -q src || docker exec -u postgres ${COMPOSE_PROJDIR}-postgres-1 /usr/bin/psql -c '\i /host/config/schema.sql'
+docker exec -ti ${COMPOSE_PROJDIR}-builder-1 /bin/bash -ic "cd /host && make build"
 cp "bin/arrow".* "./docker/arrow/"
-mv "./docker/arrow/arrow.aarch64" "./docker/arrow/arrow.arm64"
+[ -e ./docker/arrow/arrow.aarch64 ] && mv "./docker/arrow/arrow.aarch64" "./docker/arrow/arrow.arm64"
+[ -e ./docker/arrow/arrow.x86_64 ] && mv "./docker/arrow/arrow.x86_64" "./docker/arrow/arrow.amd64"
 cp config/pgarrow.yml "./docker/arrow/"
 docker-compose up -d "pgarrow${ARROW_MSGBUS}" "${ARROW_MSGBUS}arrowpg"
 exit
@@ -20,12 +21,12 @@ assert primary 'host1'
 assert primaries '[ host1 ]'
 assert standbys '[ host2, host3 ]'
 
-docker exec ${COMPOSE_PROJDIR}_postgres_2 /entrypoint.sh promote
+docker exec ${COMPOSE_PROJDIR}-postgres-2 /entrypoint.sh promote
 assert primary ''
 assert primaries '[ host1, host2 ]'
 assert standbys '[ host3 ]'
 
-docker exec ${COMPOSE_PROJDIR}_postgres_1 /entrypoint.sh rebuild
+docker exec ${COMPOSE_PROJDIR}-postgres-1 /entrypoint.sh rebuild
 assert primary 'host2'
 assert primaries '[ host2 ]'
 assert standbys '[ host1, host3 ]'
